optimized smith normal form
I figured out what was wrong with the ersion of snf that didn't transpose the matrix at every step. However, I also found a test case that breaks both snf functions in the same way that is proably inconsequential to the homology computations, but should still be fixed. Also found that findKernel isn't working properly.